moves applying-code to onApplyDefinition
